refactor: :recycle: organizes files and creates constants to generate prompt
