+++ src/modules/destinations/interfaces/ai.service.interface.ts
